Move flatten into utils
