+++ src/create-element.ts
@@ -1,15 +1,5 @@
 import { Component, Element } from './types';
-
-const flatten = <P>(
-  items: ReadonlyArray<P | ReadonlyArray<P>>
-): ReadonlyArray<P> => {
-  return items.reduce<ReadonlyArray<P>>((memo, item) => {
-    const flat = Array.isArray(item)
-      ? [...memo, ...flatten(item)]
-      : [...memo, item];
-    return flat;
-  }, []);
-};
+import { flatten } from './utils';
 
 export function createElement<P = {}>(
   type: Component<P>,
